[#936] Extract video-card-gamma/named-color/wcs tags into tags/ (item 6)
Split VideoCardGammaType/VideoCardGammaFormulaType/VideoCardGammaTableType into
tags/video_card_gamma.py, NamedColor2Value/NamedColor2ValueTuple/NamedColor2Type
into tags/named_color.py, and WcsProfilesTagType into tags/wcs.py.

NamedColor2Type.add_color() uses a deferred import for ICCProfileInvalidError,
which stays in icc_profile/__init__.py until item 10.
