+++ DisplayCAL/icc_profile/__init__.py
@@ -23,7 +23,6 @@ import struct
 import subprocess as sp
 import sys
 import warnings
-from copy import copy
 from hashlib import md5
 from typing import (
     TYPE_CHECKING,
@@ -123,16 +122,10 @@ from DisplayCAL.icc_profile.codecs import (
     s15f16_is_equal,
     s15Fixed16Number,
     s15Fixed16Number_tohex,
-    u16Fixed16Number,
-    u16Fixed16Number_tohex,
-    uInt8Number,
     uInt8Number_tohex,
-    uInt16Number,
     uInt16Number_tohex,
     uInt32Number,
     uInt32Number_tohex,
-    uInt64Number,
-    uInt64Number_tohex,
     videoCardGamma,
 )
 from DisplayCAL.icc_profile.constants import (
@@ -188,6 +181,11 @@ from DisplayCAL.icc_profile.tags.colorant import (
 from DisplayCAL.icc_profile.tags.curve import CurveType, ParametricCurveType
 from DisplayCAL.icc_profile.tags.dict_type import DictType, DictTypeJSONEncoder
 from DisplayCAL.icc_profile.tags.lut import LUT16Type
+from DisplayCAL.icc_profile.tags.named_color import (
+    NamedColor2Type,
+    NamedColor2Value,
+    NamedColor2ValueTuple,
+)
 from DisplayCAL.icc_profile.tags.text import (
     MakeAndModelType,
     MultiLocalizedUnicodeType,
@@ -195,6 +193,12 @@ from DisplayCAL.icc_profile.tags.text import (
     TextDescriptionType,
     TextType,
 )
+from DisplayCAL.icc_profile.tags.video_card_gamma import (
+    VideoCardGammaFormulaType,
+    VideoCardGammaTableType,
+    VideoCardGammaType,
+)
+from DisplayCAL.icc_profile.tags.wcs import WcsProfilesTagType
 
 
 def create_RGB_A2B_XYZ(  # noqa: N802
@@ -2961,965 +2965,6 @@ def _mp_hdr_tonemap(
     elif its_hi:
         print("Max iterations", its_hi)
     return HDR_XYZ
-
-
-class VideoCardGammaType(ICCProfileTag, ADict):
-    """Video Card Gamma Tag.
-
-    This tag contains the gamma correction values for the red, green and blue
-    channels of the video card. The values are stored in a table or as a
-    formula. The table is a 256-entry table with values ranging from 0 to
-    65535. The formula is a gamma correction formula with the following
-    parameters: redMin, redMax, redGamma, greenMin, greenMax, greenGamma,
-    blueMin, blueMax, blueGamma.
-
-    Private tag
-    http://developer.apple.com/documentation/GraphicsImaging/Reference/ColorSync_Manager/Reference/reference.html#//apple_ref/doc/uid/TP30000259-CH3g-C001473
-
-    Args:
-        tagData (bytes): The raw tag data.
-        tagSignature (str): The signature of the tag, usually "vcgt".
-    """
-
-    def __init__(self, tagData: bytes, tagSignature: str) -> None:  # noqa: N803
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-
-    def is_linear(self, r: bool = True, g: bool = True, b: bool = True) -> bool:
-        """Check if the gamma correction is linear for the red, green and blue channels.
-
-        Args:
-            r (bool): Whether to check the red channel.
-            g (bool): Whether to check the green channel.
-            b (bool): Whether to check the blue channel.
-
-        Returns:
-            bool: True if the gamma correction is linear for the specified
-                channels.
-        """
-        r_points, g_points, b_points, linear_points = self.get_values()
-        if (
-            (r and g and b and r_points == g_points == b_points)
-            or (r and g and r_points == g_points)
-            or not (g or b)
-        ):
-            points = r_points
-        elif (
-            (r and b and r_points == b_points)
-            or (g and b and g_points == b_points)
-            or not (r or g)
-        ):
-            points = b_points
-        elif g:
-            points = g_points
-        return points == linear_points
-
-    def get_unique_values(
-        self, r: bool = True, g: bool = True, b: bool = True
-    ) -> tuple:
-        """Return unique values for the red, green and blue channels.
-
-        Args:
-            r (bool): Whether to include red channel values.
-            g (bool): Whether to include green channel values.
-            b (bool): Whether to include blue channel values.
-
-        Returns:
-            tuple: Three sets containing the unique values for the red,
-        """
-        r_points, g_points, b_points, linear_points = self.get_values()
-        r_unique = {round(y) for x, y in r_points}
-        g_unique = {round(y) for x, y in g_points}
-        b_unique = {round(y) for x, y in b_points}
-        return r_unique, g_unique, b_unique
-
-    def get_values(self, r: bool = True, g: bool = True, b: bool = True) -> tuple:
-        """Return the gamma correction values for the red, green and blue channels.
-
-        Args:
-            r (bool, optional): Whether to include red channel values.
-            g (bool, optional): Whether to include green channel values.
-            b (bool, optional): Whether to include blue channel values.
-
-        Returns:
-            tuple: Four lists containing the red, green, blue, and linear
-        """
-        r_points = []
-        g_points = []
-        b_points = []
-        linear_points = []
-        vcgt = self
-        if "data" in vcgt:  # table
-            data = list(vcgt["data"])
-            while len(data) < 3:
-                data.append(data[0])
-            irange = list(range(vcgt["entryCount"]))
-            vmax = math.pow(256, vcgt["entrySize"]) - 1
-            for i in irange:
-                j = i * (255.0 / (vcgt["entryCount"] - 1))
-                linear_points.append(
-                    [j, round(i / float(vcgt["entryCount"] - 1) * 65535)]
-                )
-                if r:
-                    n = round(float(data[0][i]) / vmax * 65535)
-                    r_points.append([j, n])
-                if g:
-                    n = round(float(data[1][i]) / vmax * 65535)
-                    g_points.append([j, n])
-                if b:
-                    n = round(float(data[2][i]) / vmax * 65535)
-                    b_points.append([j, n])
-        else:  # formula
-            irange = list(range(256))
-            step = 100.0 / 255.0
-            for i in irange:
-                linear_points.append([i, i / 255.0 * 65535])
-                if r:
-                    vmin = vcgt["redMin"] * 65535
-                    v = math.pow(step * i / 100.0, vcgt["redGamma"])
-                    vmax = vcgt["redMax"] * 65535
-                    r_points.append([i, round(vmin + v * (vmax - vmin))])
-                if g:
-                    vmin = vcgt["greenMin"] * 65535
-                    v = math.pow(step * i / 100.0, vcgt["greenGamma"])
-                    vmax = vcgt["greenMax"] * 65535
-                    g_points.append([i, round(vmin + v * (vmax - vmin))])
-                if b:
-                    vmin = vcgt["blueMin"] * 65535
-                    v = math.pow(step * i / 100.0, vcgt["blueGamma"])
-                    vmax = vcgt["blueMax"] * 65535
-                    b_points.append([i, round(vmin + v * (vmax - vmin))])
-        return r_points, g_points, b_points, linear_points
-
-    def printNormalizedValues(  # noqa: N802
-        self, amount: None | int = None, digits: int = 12
-    ) -> None:
-        """Normalize and prints all values in the vcgt (range of 0.0...1.0).
-
-        For a 256-entry table with linear values from 0 to 65535:
-        #   REF            C1             C2             C3
-        001 0.000000000000 0.000000000000 0.000000000000 0.000000000000
-        002 0.003921568627 0.003921568627 0.003921568627 0.003921568627
-        003 0.007843137255 0.007843137255 0.007843137255 0.007843137255
-        ...
-        You can also specify the amount of values to print (where a value
-        lesser than the entry count will leave out intermediate values)
-        and the number of digits.
-
-        Args:
-            amount (None | int, optional): The number of values to print.
-                If None, it defaults to the entryCount if available, otherwise
-                to 256.
-            digits (int, optional): The number of digits to round the values
-                to. Defaults to 12.
-        """
-        if amount is None:
-            # use entryCount if exists, otherwise use the common value
-            amount = self.entryCount if hasattr(self, "entryCount") else 256
-        values = self.getNormalizedValues(amount)
-        entry_count = len(values)
-        channels = len(values[0])
-        header = ["REF"]
-        header.extend(f"C{k + 1}" for k in range(channels))
-        header = [title.ljust(digits + 2) for title in header]
-        print("#".ljust(len(str(amount)) + 1) + " ".join(header))
-        for i, value in enumerate(values):
-            formatted_values = [
-                str(round(channel, digits)).ljust(digits + 2, "0") for channel in value
-            ]
-            print(
-                str(i + 1).rjust(len(str(amount)), "0"),
-                str(round(i / float(entry_count - 1), digits)).ljust(digits + 2, "0"),
-                " ".join(formatted_values),
-            )
-
-
-class VideoCardGammaFormulaType(VideoCardGammaType):
-    """Video card gamma formula type class.
-
-    Args:
-        tagData (bytes): The raw tag data containing the video LUT curves.
-        tagSignature (str): The signature of the tag, typically "vcgt".
-    """
-
-    def __init__(self, tagData: bytes, tagSignature: str) -> None:  # noqa: N803
-        VideoCardGammaType.__init__(self, tagData, tagSignature)
-        data = tagData[12:]
-        self.update(
-            {
-                "redGamma": u16Fixed16Number(data[0:4]),
-                "redMin": u16Fixed16Number(data[4:8]),
-                "redMax": u16Fixed16Number(data[8:12]),
-                "greenGamma": u16Fixed16Number(data[12:16]),
-                "greenMin": u16Fixed16Number(data[16:20]),
-                "greenMax": u16Fixed16Number(data[20:24]),
-                "blueGamma": u16Fixed16Number(data[24:28]),
-                "blueMin": u16Fixed16Number(data[28:32]),
-                "blueMax": u16Fixed16Number(data[32:36]),
-            }
-        )
-
-    def getNormalizedValues(self, amount: None | int = None) -> list:  # noqa: N802
-        """Return normalized values of the video LUT curves.
-
-        Args:
-            amount (None | int, optional): The number of values to return. If
-                None, it defaults to 256.
-
-        Returns:
-            list: A list of tuples, each containing normalized values for the
-                red, green, and blue channels.
-        """
-        if amount is None:
-            amount = 256  # common value
-        step = 1.0 / float(amount - 1)
-        rgb = AODict([("red", []), ("green", []), ("blue", [])])
-        for i in range(amount):
-            for key in rgb:
-                rgb[key].append(
-                    float(self[key + "Min"])
-                    + math.pow(step * i / 1.0, float(self[key + "Gamma"]))
-                    * float(self[key + "Max"] - self[key + "Min"])
-                )
-        return list(zip(*list(rgb.values())))
-
-    def getTableType(  # noqa: N802
-        self,
-        entryCount: int = 256,  # noqa: N803
-        entrySize: int = 2,  # noqa: N803
-        quantizer: Callable = round,  # noqa: N803
-    ) -> VideoCardGammaTableType:
-        """Return gamma as table type.
-
-        Args:
-            entryCount (int, optional): The number of entries in the table.
-                Defaults to 256.
-            entrySize (int, optional): The size of each entry in bytes.
-                Defaults to 2.
-            quantizer (Callable, optional): A function to quantize the values.
-                Defaults to `round`.
-
-        Returns:
-            VideoCardGammaTableType: A new instance of VideoCardGammaTableType
-                containing the gamma table data.
-        """
-        max_value = math.pow(256, entrySize) - 1
-        tag_data = [
-            self.tagData[:8],
-            uInt32Number_tohex(0),  # type 0 = table
-            uInt16Number_tohex(3),  # channels
-            uInt16Number_tohex(entryCount),
-            uInt16Number_tohex(entrySize),
-        ]
-        int2hex = {
-            1: uInt8Number_tohex,
-            2: uInt16Number_tohex,
-            4: uInt32Number_tohex,
-            8: uInt64Number_tohex,
-        }
-        for key in ("red", "green", "blue"):
-            for i in range(entryCount):
-                vmin = float(self[key + "Min"])
-                vmax = float(self[key + "Max"])
-                gamma = float(self[key + "Gamma"])
-                v = vmin + math.pow(1.0 / (entryCount - 1) * i, gamma) * float(
-                    vmax - vmin
-                )
-                tag_data.append(int2hex[entrySize](quantizer(v * max_value)))
-        return VideoCardGammaTableType(b"".join(tag_data), self.tagSignature)
-
-
-class VideoCardGammaTableType(VideoCardGammaType):
-    """Video card gamma table type class.
-
-    Args:
-        tagData (bytes): The raw tag data containing the video LUT curves.
-        tagSignature (str): The signature of the tag, typically "vcgt".
-    """
-
-    def __init__(self, tagData: bytes, tagSignature: str) -> None:  # noqa: N803
-        VideoCardGammaType.__init__(self, tagData, tagSignature)
-        if not tagData:
-            self.update({"channels": 0, "entryCount": 0, "entrySize": 0, "data": []})
-            return
-        data = tagData[12:]
-        channels = uInt16Number(data[0:2])
-        entry_count = uInt16Number(data[2:4])
-        entry_size = uInt16Number(data[4:6])
-        self.update(
-            {
-                "channels": channels,
-                "entryCount": entry_count,
-                "entrySize": entry_size,
-                "data": [],
-            }
-        )
-        hex2int = {1: uInt8Number, 2: uInt16Number, 4: uInt32Number, 8: uInt64Number}
-        if entry_size not in hex2int:
-            raise ValueError(
-                f"Invalid VideoCardGammaTableType entry size {int(entry_size):d}"
-            )
-        i = 0
-        while i < channels:
-            self.data.append([])
-            j = 0
-            while j < entry_count:
-                index = 6 + i * entry_count * entry_size + j * entry_size
-                self.data[i].append(
-                    hex2int[entry_size](data[index : index + entry_size])
-                )
-                j = j + 1
-            i = i + 1
-
-    def getNormalizedValues(self, amount: None | int = None) -> list:  # noqa: N802
-        """Return normalized values of the video LUT curves.
-
-        Args:
-            amount (None | int, optional): The number of values to return. If
-                None, it defaults to the entryCount of the video LUT curves.
-
-        Returns:
-            list: A list of tuples, each containing normalized values for the
-                red, green, and blue channels.
-        """
-        if amount is None:
-            amount = self.entryCount
-        max_value = math.pow(256, self.entrySize) - 1
-        values = list(
-            zip(*[[entry / max_value for entry in channel] for channel in self.data])
-        )
-        if amount <= self.entryCount:
-            step = self.entryCount / float(amount - 1)
-            all_values = values
-            values = []
-            for i, value in enumerate(all_values):
-                if i == 0 or (i + 1) % step < 1 or i + 1 == self.entryCount:
-                    values.append(value)
-        return values
-
-    def getFormulaType(self) -> VideoCardGammaFormulaType:  # noqa: N802
-        """Return formula representing gamma value at 50% input.
-
-        Returns:
-            VideoCardGammaFormulaType: A new instance of
-                VideoCardGammaFormulaType with the calculated gamma values and
-                min/max values for each channel.
-        """
-        max_value = math.pow(256, self.entrySize) - 1
-        tag_data = [self.tagData[:8], uInt32Number_tohex(1)]  # type 1 = formula
-        data = list(self.data)
-        while len(data) < 3:
-            data.append(data[0])
-        for channel in data:
-            channel_length = (len(channel) - 1) / 2.0
-            floor = float(channel[math.floor(channel_length)])
-            ceil = float(channel[math.ceil(channel_length)])
-            vmin = channel[0] / max_value
-            vmax = channel[-1] / max_value
-            v = (vmin + ((floor + ceil) / 2.0) * (vmax - vmin)) / max_value
-            gamma = math.log(v) / math.log(0.5)
-            print(vmin, gamma, vmax)
-            tag_data.append(u16Fixed16Number_tohex(gamma))
-            tag_data.append(u16Fixed16Number_tohex(vmin))
-            tag_data.append(u16Fixed16Number_tohex(vmax))
-        return VideoCardGammaFormulaType(b"".join(tag_data), self.tagSignature)
-
-    def quantize(self, bits: int = 16, quantizer: Callable = round) -> None:
-        """Quantize to n bits of precision.
-
-        Note that when the quantize bits are not 8, 16, 32 or 64, double
-        quantization will occur: First from the table precision bits according
-        to entrySize to the chosen quantization bits, and then back to the
-        table precision bits.
-
-        Args:
-            bits (int, optional): The number of bits to quantize to. Must be
-                one of 8, 16, 32, or 64. Defaults to 16.
-            quantizer (callable, optional): A function to quantize the values.
-                Defaults to the built-in `round` function.
-        """
-        oldmax = math.pow(256, self.entrySize) - 1
-        if bits in (8, 16, 32, 64):
-            self.entrySize = int(bits / 8)
-        bitv = 2.0**bits
-        newmax = math.pow(256, self.entrySize) - 1
-        for _i, channel in enumerate(self.data):
-            for j, value in enumerate(channel):
-                channel[j] = int(quantizer(value / oldmax * bitv) / bitv * newmax)
-
-    def resize(self, length: int = 128) -> None:
-        """Resize video LUT curves to a given length.
-
-        Args:
-            length (int): The desired length of the resized LUT curves.
-        """
-        data = [[], [], []]
-        for i, channel in enumerate(self.data):
-            for j in range(length):
-                j *= (len(channel) - 1) / float(length - 1)
-                if int(j) != j:
-                    floor = channel[math.floor(j)]
-                    ceil = channel[min(math.ceil(j), len(channel) - 1)]
-                    interpolated = range(floor, ceil + 1)
-                    fraction = j - int(j)
-                    index = round(fraction * (ceil - floor))
-                    v = interpolated[index]
-                else:
-                    v = channel[int(j)]
-                data[i].append(v)
-        self.data = data
-        self.entryCount = len(data[0])
-
-    def resized(self, length: int = 128) -> VideoCardGammaTableType:
-        """Return a resized version of the video LUT curves.
-
-        Args:
-            length (int): The desired length of the resized LUT curves.
-
-        Returns:
-            VideoCardGammaTableType: A new instance of VideoCardGammaTableType
-                with the resized LUT curves.
-        """
-        resized = self.__class__(self.tagData, self.tagSignature)
-        resized.resize(length)
-        return resized
-
-    def smooth_cr(self, length: int = 64) -> None:
-        """Smooth video LUT curves (Catmull-Rom).
-
-        Args:
-            length (int): The desired length of the smoothed LUT curves.
-                Defaults to 64.
-        """
-        resized = self.resized(length)
-        for i in range(len(self.data)):
-            step = float(length - 1) / (len(self.data[i]) - 1)
-            interpolation = CRInterpolation(resized.data[i])
-            for j in range(len(self.data[i])):
-                self.data[i][j] = interpolation(j * step)
-
-    def smooth_avg(self, passes: int = 1, window: None | list | tuple = None) -> None:
-        """Smooth video LUT curves (moving average).
-
-        Args:
-            passes (int): Number of passes to perform. Defaults to 1.
-            window (None | list | tuple , optional): Tuple or list containing
-                weighting factors. Its length determines the size of the window
-                to use. Defaults to (1.0, 1.0, 1.0).
-        """
-        for i, channel in enumerate(self.data):
-            self.data[i] = colormath.smooth_avg(channel, passes, window)
-        self.entryCount = len(self.data[0])
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: The raw tag data formatted as bytes.
-        """
-        tag_data = [
-            b"vcgt",
-            b"\0" * 4,
-            uInt32Number_tohex(0),  # type 0 = table
-            uInt16Number_tohex(len(self.data)),  # channels
-            uInt16Number_tohex(self.entryCount),
-            uInt16Number_tohex(self.entrySize),
-        ]
-        int2hex = {
-            1: uInt8Number_tohex,
-            2: uInt16Number_tohex,
-            4: uInt32Number_tohex,
-            8: uInt64Number_tohex,
-        }
-        tag_data.extend(
-            int2hex[self.entrySize](channel[i])
-            for channel in self.data
-            for i in range(self.entryCount)
-        )
-        return b"".join(tag_data)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        """Set the tag data.
-
-        Does nothing in this case, as the tagData is generated
-        from the internal data structure.
-
-        Args:
-            tagData (bytes): The raw tag data to set.
-        """
-
-
-class WcsProfilesTagType(ICCProfileTag, ADict):
-    """ICC WCS profiles tag type.
-
-    Args:
-        tagData (bytes): The raw tag data.
-        tagSignature (str): The signature of the tag.
-        profile (ICCProfile): The ICC profile to which this tag belongs.
-    """
-
-    def __init__(self, tagData: bytes, tagSignature: str, profile: ICCProfile) -> None:  # noqa: N803
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        self.profile = profile
-        for i, modelname in enumerate(
-            ["ColorDeviceModel", "ColorAppearanceModel", "GamutMapModel"]
-        ):
-            j = i * 8
-            if len(tagData) < 16 + j:
-                break
-            offset = uInt32Number(tagData[8 + j : 12 + j])
-            size = uInt32Number(tagData[12 + j : 16 + j])
-            if offset and size:
-                from io import StringIO
-
-                from defusedxml import ElementTree
-
-                it = ElementTree.iterparse(StringIO(tagData[offset : offset + size]))
-                for _event, elem in it:
-                    elem.tag = elem.tag.split("}", 1)[-1]  # Strip all namespaces
-                self[modelname] = it.root
-
-    def get_vcgt(
-        self,
-        quantize: int | bool = False,
-        quantizer: Callable = round,
-    ) -> None | VideoCardGammaType:
-        """Return calibration information (if present) as VideoCardGammaType.
-
-        If quantize is set, a table quantized to <quantize> bits is returned.
-
-        Note that when the quantize bits are not 8, 16, 32 or 64, multiple
-        quantizations will occur: For quantization bits below 32, first to 32
-        bits, then to the chosen quantization bits, then back to 32 bits (which
-        will be the final table precision bits).
-
-        Args:
-            quantize (bool | int, optional): If True, quantize to 16 bits
-                (default). If an integer, quantize to that many bits.
-            quantizer (Callable, optional): A quantization function, defaults to
-                `round`.
-
-        Returns:
-            None | VideoCardGammaType: Returns a VideoCardGammaType object if
-                calibration information is present, otherwise None.
-        """
-        if quantize and not isinstance(quantize, int):
-            raise ValueError(f"Invalid quantization bits: {quantize!r}")
-
-        if "ColorDeviceModel" not in self:
-            return None
-
-        # Parse calibration information to VCGT
-        cal = self.ColorDeviceModel.find("Calibration")
-        if cal is None:
-            return None
-        agammaconf = cal.find("AdapterGammaConfiguration")
-        if agammaconf is None:
-            return None
-        pcurves = agammaconf.find("ParameterizedCurves")
-        if pcurves is None:
-            return None
-        vcgt_data = "vcgt"
-        vcgt_data += b"\0" * 4
-        vcgt_data += uInt32Number_tohex(1)  # Type 1 = formula
-        for color in ("Red", "Green", "Blue"):
-            trc = pcurves.find(color + "TRC")
-            if trc is None:
-                trc = {}
-            vcgt_data += u16Fixed16Number_tohex(float(trc.get("Gamma", 1)))
-            vcgt_data += u16Fixed16Number_tohex(float(trc.get("Offset1", 0)))
-            vcgt_data += u16Fixed16Number_tohex(float(trc.get("Gain", 1)))
-        vcgt = VideoCardGammaFormulaType(vcgt_data, "vcgt")
-        if quantize:
-            if quantize in (8, 16, 32, 64):
-                entry_size = quantize / 8
-            elif quantize < 32:
-                entry_size = 4
-            else:
-                entry_size = 8
-            vcgt = vcgt.getTableType(entrySize=entry_size, quantizer=quantizer)
-            if quantize not in (8, 16, 32, 64):
-                vcgt.quantize(quantize, quantizer)
-        return vcgt
-
-
-class NamedColor2Value:
-    """Named Color 2 Value.
-
-    Args:
-        valueData (bytes, optional): Binary data containing the named color
-            values.
-        deviceCoordCount (int, optional): Number of device coordinates.
-        pcs (str, optional): PCS name, either "XYZ" or "Lab".
-        device (str, optional): Device name, either "RGB" or "Lab".
-    """
-
-    def __init__(
-        self,
-        valueData: bytes = b"\0" * 38,  # noqa: N803
-        deviceCoordCount: int = 0,  # noqa: N803
-        pcs: str = "XYZ",
-        device: str = "RGB",
-    ) -> None:
-        self._pcsname = pcs
-        self._devicename = device
-        end = valueData[0:32].find(b"\0")
-        if end < 0:
-            end = 32
-        self.rootName = valueData[0:end]
-        self.pcsvalues = [
-            uInt16Number(valueData[32:34]),
-            uInt16Number(valueData[34:36]),
-            uInt16Number(valueData[36:38]),
-        ]
-
-        self.pcs = AODict()
-        for i, pcsvalue in enumerate(self.pcsvalues):
-            if pcs == "Lab":
-                if i == 0:
-                    # L* range 0..100 + (25500 / 65280.0)
-                    self.pcs[pcs[i]] = pcsvalue / 65536.0 * 256 / 255.0 * 100
-                else:
-                    # a, b range -128..127 + (255/256.0)
-                    self.pcs[pcs[i]] = -128 + (pcsvalue / 65536.0 * 256)
-            elif pcs == "XYZ":
-                # X, Y, Z range 0..100 + (32767 / 32768.0)
-                self.pcs[pcs[i]] = pcsvalue / 32768.0 * 100
-
-        device_coords = []
-        if deviceCoordCount > 0:
-            device_coords.extend(
-                uInt16Number(valueData[i : i + 2])
-                for i in range(38, 38 + deviceCoordCount * 2, 2)
-            )
-        self.devicevalues = device_coords
-        if device == "Lab":
-            # L* range 0..100 + (25500 / 65280.0)
-            # a, b range range -128..127 + (255 / 256.0)
-            self.device = tuple(
-                (
-                    v / 65536.0 * 256 / 255.0 * 100
-                    if i == 0
-                    else -128 + (v / 65536.0 * 256)
-                )
-                for i, v in enumerate(device_coords)
-            )
-        elif device == "XYZ":
-            # X, Y, Z range 0..100 + (32767 / 32768.0)
-            self.device = tuple(v / 32768.0 * 100 for v in device_coords)
-        else:
-            # Device range 0..100
-            self.device = tuple(v / 65535.0 * 100 for v in device_coords)
-
-    @property
-    def name(self) -> str:
-        """Return the name of the named color.
-
-        Returns:
-            str: The name of the named color, decoded from bytes using
-                'latin-1' encoding.
-        """
-        return str(Text(self.rootName.strip(b"\0")), "latin-1")
-
-    def __repr__(self) -> str:
-        """Return the string representation of the object.
-
-        Returns:
-            str: The string representation of the object.
-        """
-        pcs = []
-        for key in self.pcs:
-            value = self.pcs[key]
-            pcs.append(f"{key}={value}")
-        dev = [f"{value}" for value in self.device]
-        return "{}({}, {{{}}}, [{}])".format(
-            self.__class__.__name__,
-            self.name,
-            ", ".join(pcs),
-            ", ".join(dev),
-        )
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: Raw tag data containing the named color values.
-        """
-        value_data = []
-        value_data.append(self.rootName.ljust(32, b"\0"))
-        value_data.extend([uInt16Number_tohex(pcsval) for pcsval in self.pcsvalues])
-        value_data.extend(
-            [uInt16Number_tohex(deviceval) for deviceval in self.devicevalues]
-        )
-        return b"".join(value_data)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        """Set tag data.
-
-        Does nothing, as NamedColor2Value is immutable.
-
-        Args:
-            tagData (bytes): Raw tag data to set.
-        """
-
-
-class NamedColor2ValueTuple(tuple):
-    """Tuple subclass for NamedColor2Value.
-
-    This class is used to represent a tuple of NamedColor2Value objects.
-    """
-
-    __slots__ = ()
-    REPR_OUTPUT_SIZE = 10
-
-    def __repr__(self) -> str:
-        """Return the string representation of the object.
-
-        Truncates the output if it exceeds the specified size.
-
-        Returns:
-            str: The string representation of the object.
-        """
-        data = list(self[: self.REPR_OUTPUT_SIZE + 1])
-        if len(data) > self.REPR_OUTPUT_SIZE:
-            data[-1] = "...(remaining elements truncated)..."
-        return repr(data)
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: Concatenated tag data from all NamedColor2Value objects in
-                the tuple.
-        """
-        return b"".join([val.tagData for val in self])
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        pass
-
-
-class NamedColor2Type(ICCProfileTag, AODict):
-    """Named Color 2 Type.
-
-    This tag contains a list of named colors, each with a set of device
-    coordinates and a set of PCS coordinates. The device coordinates
-    are used to identify the color on the device, while the PCS
-    coordinates are used to identify the color in a device-independent
-    color space. The tag also contains a prefix and suffix that are
-    used to format the name of the color.
-
-    Byte offset content encoded as:
-
-        0..3    vendorData        s4Fixed32Number
-        4..7    colorCount        uInt32Number
-        8..11   deviceCoordCount  uInt32Number
-        12..15  reserved          uInt32Number
-        16..19  reserved          uInt32Number
-        20..51  prefix            s32Fixed32Number
-        52..83  suffix            s32Fixed32Number
-        84..n   colorValues       NamedColor2Value
-
-    Args:
-        tagData (bytes): The raw tag data.
-        tagSignature (None | str): The signature of the tag.
-        pcs (None | str): The PCS name, either "XYZ" or "Lab".
-        device (None | str): The device name, either "RGB" or "Lab".
-    """
-
-    REPR_OUTPUT_SIZE = 10
-
-    def __init__(
-        self,
-        tagData: bytes = b"\0" * 84,  # noqa: N803
-        tagSignature: None | str = None,  # noqa: N803
-        pcs: None | str = None,
-        device: None | str = None,
-    ) -> None:
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        AODict.__init__(self)
-
-        colorCount = uInt32Number(tagData[12:16])  # noqa: N806
-        deviceCoordCount = uInt32Number(tagData[16:20])  # noqa: N806
-        stride = 38 + 2 * deviceCoordCount
-
-        self.vendorData = tagData[8:12]
-        self.colorCount = colorCount
-        self.deviceCoordCount = deviceCoordCount
-        self._prefix = Text(tagData[20:52])
-        self._suffix = Text(tagData[52:84])
-        self._pcsname = pcs
-        self._devicename = device
-
-        keys = []
-        values = []
-        if colorCount > 0:
-            start = 84
-            end = start + (stride * colorCount)
-            for i in range(start, end, stride):
-                nc2 = NamedColor2Value(
-                    tagData[i : i + stride], deviceCoordCount, pcs=pcs, device=device
-                )
-                keys.append(nc2.name)
-                values.append(nc2)
-        self.update(dict(list(zip(keys, values))))
-
-    def __setattr__(self, name: str, value: Any) -> None:  # noqa: ANN401
-        """Set an attribute of the object.
-
-        Args:
-            name (str): The name of the attribute to set.
-            value (Any): The value to set the attribute to.
-        """
-        object.__setattr__(self, name, value)
-
-    @property
-    def prefix(self) -> str:
-        """Return the prefix of the named color profile.
-
-        Returns:
-            str: The prefix of the named color profile, decoded from bytes
-                using 'latin-1' encoding.
-        """
-        return str(self._prefix.strip(b"\0"), "latin-1")
-
-    @property
-    def suffix(self) -> str:
-        """Return the suffix of the named color profile.
-
-        Returns:
-            str: The suffix of the named color profile, decoded from bytes
-                using 'latin-1' encoding.
-        """
-        return str(self._suffix.strip(b"\0"), "latin-1")
-
-    @property
-    def colorValues(self) -> NamedColor2ValueTuple:  # noqa: N802
-        """Return a tuple of NamedColor2Value objects.
-
-        Returns:
-            NamedColor2ValueTuple: A tuple containing all NamedColor2Value
-                objects in the profile.
-        """
-        return NamedColor2ValueTuple(list(self.values()))
-
-    def add_color(
-        self,
-        root_name: str,
-        *device_coordinates: list[float],
-        **pcs_coordinates: dict[str, float],
-    ) -> None:
-        """Add a named color to the profile.
-
-        Args:
-            root_name (str): The name of the color.
-            device_coordinates (list): Device coordinates for the color.
-            pcs_coordinates (dict): PCS coordinates for the color.
-
-        Raises:
-            ICCProfileInvalidError: If the required PCS coordinates or device
-                coordinates are not provided, or if the color name already
-                exists.
-        """
-        if self._pcsname == "Lab":
-            keys = ["L", "a", "b"]
-        elif self._pcsname == "XYZ":
-            keys = ["X", "Y", "Z"]
-        else:
-            keys = ["X", "Y", "Z"]
-
-        if not set(pcs_coordinates.keys()).issuperset(set(keys)):
-            raise ICCProfileInvalidError(
-                "Can't add namedColor2 without all 3 PCS coordinates: "  # noqa: UP032
-                "'{}'".format(set(keys) - set(pcs_coordinates.keys()))
-            )
-
-        if len(device_coordinates) != self.deviceCoordCount:
-            raise ICCProfileInvalidError(
-                f"Can't add namedColor2 without all {self.deviceCoordCount} "
-                f"device coordinates (called with {len(device_coordinates)})"
-            )
-
-        nc2value = NamedColor2Value()
-        nc2value._pcsname = self._pcsname
-        nc2value._devicename = self._devicename
-        nc2value.rootName = root_name
-
-        if root_name in list(self.keys()):
-            raise ICCProfileInvalidError(
-                f"Can't add namedColor2 with existant name: '{root_name}'"
-            )
-
-        nc2value.devicevalues = []
-        nc2value.device = tuple(device_coordinates)
-        nc2value.pcs = AODict(copy(pcs_coordinates))
-
-        for idx, key in enumerate(keys):
-            val = nc2value.pcs[key]
-            if key == "L":
-                nc2value.pcsvalues[idx] = val * 65536 / (256 / 255.0) / 100.0
-            elif key in ("a", "b"):
-                nc2value.pcsvalues[idx] = (val + 128) * 65536 / 256.0
-            elif key in ("X", "Y", "Z"):
-                nc2value.pcsvalues[idx] = val * 32768 / 100.0
-
-        for idx, val in enumerate(nc2value.device):
-            if self._devicename == "Lab":
-                if idx == 0:
-                    # L* range 0..100 + (25500 / 65280.0)
-                    nc2value.devicevalues[idx] = val * 65536 / (256 / 255.0) / 100.0
-                else:
-                    # a, b range -128..127 + (255/256.0)
-                    nc2value.devicevalues[idx] = (val + 128) * 65536 / 256.0
-            elif self._devicename == "XYZ":
-                # X, Y. Z range 0..100 + (32767 / 32768.0)
-                nc2value.devicevalues[idx] = val * 32768 / 100.0
-            else:
-                # Device range 0..100
-                nc2value.devicevalues[idx] = val * 65535 / 100.0
-
-        self[nc2value.name] = nc2value
-
-    def __repr__(self) -> str:
-        """Return the string representation of the object.
-
-        Truncates the output if it exceeds the specified size.
-
-        Returns:
-            str: The string representation of the object.
-        """
-        data = list(self.items())[: self.REPR_OUTPUT_SIZE + 1]
-        if len(data) > self.REPR_OUTPUT_SIZE:
-            data[-1] = ("...", "(remaining elements truncated)")
-        return repr(dict(data))
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: Raw tag data containing vendor data, color count,
-                device coordinate count, prefix, suffix, and color values.
-        """
-        tagData = [  # noqa: N806
-            b"ncl2",
-            b"\0" * 4,
-            self.vendorData,
-            uInt32Number_tohex(len(list(self.items()))),
-            uInt32Number_tohex(self.deviceCoordCount),
-            self._prefix.ljust(32),
-            self._suffix.ljust(32),
-            self.colorValues.tagData,
-        ]
-        return b"".join(tagData)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        pass
 
 
 TAG_SIGNATURE_TO_TAG = {"arts": ChromaticAdaptionTag, "chad": ChromaticAdaptionTag}
